requirements.txt
customtkinter==5.1.2
pyperclip==1.8.2
random_password_generator==2.2.0